Eshop/requirements.txt
django>=2
pymysql
djangorestframework
coreapi
markdown
django-filter
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools==2.1
future==0.15.2
httplib2==0.9.2
six==1.10.0
xlwt
xlsxwriter
pillow
django-cors-headers
djangorestframework-jwt
pycryptodome
drf-extensions
django-redis
requests
social-auth-app-django
# pip install https://codeload.github.com/sshwsfc/xadmin/zip/django2